Update grub.sh
Now rejects running on non-Darwin platforms since this is kind of a hack.

It first installs mpfr and libmpc from Homebrew, then builds an entire
GCC/Binutils cross-compile toolchain for x86_64-elf.

It also builds objconv if requested since that is a prerequisite for
building grub.

Finally, using our new toolchain, we build grub and install it. The
script then removes mpfr and libmpc since they are unneeded, and exits