newlib: Write syscall in assembly
Inline assembly for x86_64 is a mess, since we can't explicitly state r8 as a
register. While we can work around by using register with a variable, that
seems like a hack.

Additionally, the previous inline assembly had the arguments passed in reverse,
which meant that the last argument (r9) was stored as the first (rdi), etc.