vscode: Use our toolchain for Intellisense parsing
This removes need to manually add compiler includes to browse and includePath.
While that currently works, it's inherently tied to a specific compiler
version. When updating that toolchain, we would need to update that to.

By using the cross-compiler toolchain as the parser instead of system compiler,
the compiler-specific headers are automatically included.

We also need to add -fno-exceptions to prevent GCC from complaining about how
we violated exception specifications, which is nonsense because we don't even
use exceptions.