toolchain: Patch grub to support quark toolchain
Somehow Pepper was a valid operating system that was accepted, which meant
that we were probably building with the wrong configuration too.

When we switched to $(ARCH)-quark- as the target triplet this showed up
as an error about invalid system configuration.

This commit switches grub to be built like GCC and binutils, meaning
that we now download a tarball and apply a patch to get it to build.